feat: 참여자 역할 생성, 플레이어와 딜러 분리
