+++ src/main/kotlin/blackjack/domain/Dealer.kt
@@ -1,0 +1,6 @@
+package blackjack.domain
+
+class Dealer(
+    override val name: String = "딜러",
+    override val hands: Hands = Hands(),
+) : Participant(name, hands)
