+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,35 +1,10 @@
 package blackjack.domain
 
 class Player(
-    val name: String,
-    val hands: Hands = Hands(),
-) {
-    val score: Int
-        get() = hands.calculateTotalValue()
-
-    val handSize: Int
-        get() = hands.size
-
-    var status: PlayerStatus = PlayerStatus.PLAYING
-        private set
-
-    fun isPlayable(): Boolean = status.isPlayable()
-
+    override val name: String,
+    override val hands: Hands = Hands(),
+) : Participant(name, hands) {
     fun toStay() {
-        this.status = PlayerStatus.STAY
-    }
-
-    fun hit(deck: Deck) {
-        if (!isPlayable()) {
-            return
-        }
-
-        val card = deck.draw()
-        hands.add(card)
-        handleStatus()
-    }
-
-    private fun handleStatus() {
-        status = status.handleStatus(score)
+        status = PlayerStatus.STAY
     }
 }
